kafka api changes
