+++ examples/README.md
@@ -18,7 +18,7 @@ Current examples:
   runtime values in an example plugin,
 * **[kafka lib](kafka_lib)** shows how to use the Kafka messaging library
   on set of individual tools (producer, sync and async consumer),
-* **[kafka plugin](kafka_plugin/main.go)** contains a simple plugin which 
+* **[kafka plugin](kafka_plugin/non_clustered/main.go)** contains a simple plugin which 
   registers a Kafka consumer and sends a test notification,
 * **[logs lib](logs_lib)** shows how to use the logger library and switch 
   between the log levels,
