+++ messaging/messaging_api.go
@@ -41,6 +41,7 @@ type ProtoWatcher interface {
 type ProtoMessage interface {
 	keyval.ProtoKvPair
 	GetTopic() string
+	GetPartition() int32
 }
 
 // ProtoMessageErr represents a message that was not published successfully to a messaging system.
